# dia_01/cu-01/openapi.json
{"openapi":"3.1.0","info":{"title":"CU-01 Azure Service Recommender API","description":"Backend del agente watsonx Orchestrate para recomendación de servicios Azure. Implementa las fases TRIAGE, INTERVIEW y HANDOFF, con fallback semántico para respuestas ambiguas y telemetría para análisis post-workshop.","version":"2.0.0"},"servers":[{"url":"https://api-decision-tree.28fzvauz0vhk.us-south.codeengine.appdomain.cloud"}],"paths":{"/metadata":{"get":{"tags":["meta"],"summary":"version_decision_tree","description":"Returns version, hash, timestamp and counts from the decision tree loaded in Cloudant. Helpfull to identify the exact versión active and log which decision tree reponse to each session.","operationId":"get_tree_metadata","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metadata"}}}}}}},"/node/start":{"get":{"tags":["triage"],"summary":"get_start_node","description":"Returns the root node of the decision tree. MANDATORY: Call this at the start of every conversation. This is part of a deterministic flow. You must not stop until the next node is retrieved.Do not ask the user anything before calling this.","operationId":"get_decision_tree_start","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/node/{node_id}":{"get":{"tags":["triage"],"summary":"get_next_node","description":"Returns the next node in the decision tree. MANDATORY: After selecting an option, you MUST call this using the `next` field. This is the ONLY valid way to continue the flow. Do not stop. Do not call other tools instead. Always continue the tree until a deep_dive or result node is reached.This is part of a deterministic flow. You must not stop until the next node is retrieved.","operationId":"get_decision_node","parameters":[{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/services":{"get":{"tags":["lookup"],"summary":"list_services","description":"Returns all available Azure services. Use ONLY if the FIRST user message explicitly names a service. FORBIDDEN during decision tree traversal. Do not use after calling the start node.","operationId":"list_services","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceList"}}}}}}},"/services/{deep_dive_id}/questionnaire":{"get":{"tags":["interview"],"summary":"get_questionnaire","description":"Returns the full questionnaire for a service. Use ONLY in INTERVIEW phase. Call immediately after reaching a deep_dive node. Do not use during decision tree traversal.","operationId":"get_service_questionnaire","parameters":[{"name":"deep_dive_id","in":"path","required":true,"schema":{"type":"string","title":"Deep Dive Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommendation":{"post":{"tags":["handoff"],"summary":"build_recommendation","description":"Builds the final recommendation. Use ONLY after completing the interview. Call exactly once per conversation. Do not use during decision tree traversal.","operationId":"build_recommendation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recommendation"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/node/{node_id}/match":{"post":{"tags":["triage"],"summary":"match_user_answer","description":"Matches a free-text answer to one of the available options. Use ONLY if the answer does not clearly match a number or label. MANDATORY: After this call, you MUST immediately call the next node using matched.next. Do not stop. Do not ask the user yet. Continue the flow.","operationId":"match_user_answer","parameters":[{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/events":{"post":{"tags":["telemetry"],"summary":"log_conversation_event","description":"Logs a conversation step for analytics. This is fire-and-forget. MANDATORY: This must be the LAST action in the step. Call this only AFTER all other tool calls are completed. Never call this before getting the next node.","operationId":"log_event","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EventRequest":{"properties":{"session_id":{"type":"string","title":"Session Id","description":"ID único de la sesión conversacional."},"phase":{"type":"string","title":"Phase","description":"triage | interview | handoff"},"node_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Node Id"},"user_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Answer"},"matched_option":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Matched Option"},"extra":{"additionalProperties":true,"type":"object","title":"Extra"}},"type":"object","required":["session_id","phase"],"title":"EventRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MatchCandidate":{"properties":{"index":{"type":"integer","title":"Index"},"label":{"type":"string","title":"Label"},"next":{"type":"string","title":"Next"},"score":{"type":"number","title":"Score"}},"type":"object","required":["index","label","next","score"],"title":"MatchCandidate"},"MatchRequest":{"properties":{"user_answer":{"type":"string","title":"User Answer","description":"Respuesta libre del usuario a una pregunta del árbol."}},"type":"object","required":["user_answer"],"title":"MatchRequest"},"MatchResponse":{"properties":{"matched":{"anyOf":[{"$ref":"#/components/schemas/MatchCandidate"},{"type":"null"}]},"confidence":{"type":"number","title":"Confidence"},"strategy":{"type":"string","title":"Strategy","description":"numeric | substring | yesno | keyword | fuzzy | none"},"candidates":{"items":{"$ref":"#/components/schemas/MatchCandidate"},"type":"array","title":"Candidates"}},"type":"object","required":["confidence","strategy","candidates"],"title":"MatchResponse"},"Metadata":{"properties":{"version":{"type":"string","title":"Version"},"hash":{"type":"string","title":"Hash"},"loaded_at":{"type":"string","title":"Loaded At"},"node_count":{"type":"integer","title":"Node Count"},"counts":{"additionalProperties":{"type":"integer"},"type":"object","title":"Counts"}},"type":"object","required":["version","hash","loaded_at","node_count","counts"],"title":"Metadata"},"Recommendation":{"properties":{"service":{"type":"string","title":"Service"},"summary":{"type":"string","title":"Summary"},"why_factors":{"items":{"type":"string"},"type":"array","title":"Why Factors"},"key_considerations":{"items":{"type":"string"},"type":"array","title":"Key Considerations"},"pricing":{"additionalProperties":true,"type":"object","title":"Pricing"},"docs_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Docs Url"},"user_context":{"additionalProperties":true,"type":"object","title":"User Context"},"handoff":{"additionalProperties":true,"type":"object","title":"Handoff"}},"type":"object","required":["service","summary"],"title":"Recommendation"},"RecommendationRequest":{"properties":{"service_id":{"type":"string","title":"Service Id","description":"_id del nodo result (p. ej. 'result_container_apps')."},"tree_path":{"items":{"type":"string"},"type":"array","title":"Tree Path","description":"IDs de los nodos recorridos en orden."},"deep_dive_answers":{"additionalProperties":{"type":"string"},"type":"object","title":"Deep Dive Answers","description":"Mapa question_id -> respuesta del usuario acumulado durante INTERVIEW."}},"type":"object","required":["service_id"],"title":"RecommendationRequest"},"ServiceList":{"properties":{"services":{"items":{"$ref":"#/components/schemas/ServiceSummary"},"type":"array","title":"Services"}},"type":"object","required":["services"],"title":"ServiceList"},"ServiceSummary":{"properties":{"id":{"type":"string","title":"Id"},"service":{"type":"string","title":"Service"},"summary":{"type":"string","title":"Summary"},"docs_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Docs Url"}},"type":"object","required":["id","service","summary"],"title":"ServiceSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}